+++ src/main/java/org/etmetmy/bn_server/domain/user/dto/MemberUpdateRequest.java
@@ -1,10 +1,12 @@
 package org.etmetmy.bn_server.domain.user.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
+@AllArgsConstructor
 public class MemberUpdateRequest {
     private String name;
     private String email;
