+++ src/main/java/org/etmetmy/bn_server/domain/user/repository/UserRepository.java
@@ -5,10 +5,12 @@ import org.etmetmy.bn_server.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param; // Param import
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 
+@Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 
     boolean existsByEmail(String email);
